+++ ui/components/_base/header/MainNav.js
@@ -5,9 +5,6 @@ import MobileNav from './MobileNav';
 import DesktopNav from './DesktopNav';
 import { useRouter } from 'next/router';
 import AllInclusiveIcon from '@mui/icons-material/AllInclusive';
-import FeedIcon from '@mui/icons-material/Feed';
-import MenuBookIcon from '@mui/icons-material/MenuBook';
-import SwapHorizIcon from '@mui/icons-material/SwapHoriz';
 import PortfolioLink from '../PortfolioLink';
 import WatchlistLink from '../WatchlistLink';
 import CurrencySelector from '../CurrencySelector';
@@ -15,6 +12,7 @@ import { DarkMode, InsertDriveFile, PieChart, Visibility } from '@mui/icons-mate
 import LightMode from '@mui/icons-material/LightMode';
 import { GeneralContext } from '../../../../contexts/general/General.Context';
 import Link from 'next/link';
+import Search from './Search';
 
 const MainNav = () => {
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
@@ -119,33 +117,36 @@ const MainNav = () => {
   ];  
 
   return (
-    <nav className='flex items-center w-full justify-end lg:justify-between'>
+    <nav className='flex items-center w-full justify-between'>
       <div className='flex items-center w-full sm:w-auto'>
         <DesktopNav navLinks={navConfig} path={path} />
         {/* Bubbles button */}
         <div className='font-semibold'>
-          <Link href="/bubbles" className={`bubbles-button py-4 px-4 block sm:hidden md:block sm:hover:text-primary relative sm:mr-[70px]`}>
+          <Link href="/bubbles" className={`bubbles-button py-4 px-4 block sm:hidden md:block sm:hover:text-primary relative`}>
             Bubbles <span className='uppercase rounded-3xl bg-orange text-black dark:text-dark-bg absolute top-2 sm:left-[87%] left-[85%] text-[10px] font-extrabold px-2 py-[2px] scale-[0.85] sm:scale-90'>Hot</span>
           </Link>
         </div>
       </div>
 
-      <div></div>
-      <div>
-        <div className='inline-block'>
-          <div className='hidden sm:inline-block'>
-            <span className='mr-4'><WatchlistLink /></span>
-            <PortfolioLink />
+      <div className='flex items-center'>
+        <div></div>
+        <div className='flex items-center'>
+          <div className='inline-block'>
+            <div>
+              <span className='sm:mr-4'><Search /></span>
+              <span className='mr-4 hidden sm:inline-block'><WatchlistLink /></span>
+              <span className='hidden sm:inline-block'><PortfolioLink /></span>
+            </div>
+          </div>
+          <div className='sm:mx-4 mr-4 inline-block'>
+            <CurrencySelector />
+          </div>
+          <div className='hidden lg:inline-block'>
+            <AccountNav />
           </div>
         </div>
-        <div className='mx-4 inline-block'>
-          <CurrencySelector />
-        </div>
-        <div className='hidden lg:inline-block'>
-          <AccountNav />
-        </div>
+        <MobileNav navLinks={navConfig} isMobileMenuOpen={isMobileMenuOpen} toggleMobileMenu={toggleMobileMenu} path={path} />
       </div>
-      <MobileNav navLinks={navConfig} isMobileMenuOpen={isMobileMenuOpen} toggleMobileMenu={toggleMobileMenu} path={path} />
     </nav>
   );
 };